fix: Update logout command to clear project and org from storage
