+++ src/utils/index.ts
@@ -1,9 +1,11 @@
 import * as loadRepoConfig from './loadRepoConfig'
 import * as showDebugOutput from './showDebugOutput'
 import * as loginAndRefresh from './loginAndRefresh'
+import * as logout from './logout'
 
 export default {
   ...loadRepoConfig,
   ...showDebugOutput,
-  ...loginAndRefresh
+  ...loginAndRefresh,
+  ...logout
 }
